xhr/send-redirect.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - Redirects (basics)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#infrastructure-for-the-send()-method" data-tested-assertations="following::dl[1]/dt[2] following::dl[1]/dd[2]/ol/li[1] following::dl[1]/dd[2]/ol/li[2]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      function redirect(code, method="GET", redirectLocation="content.py") {
        var test = async_test(`${document.title} (${code}, ${method}, ${redirectLocation})`)
        test.step(function() {
          var client = new XMLHttpRequest()
          client.onreadystatechange = function() {
            test.step(function() {
              if(client.readyState == 4) {
                let expected_method = method;
                if (method == "POST" && (code == "301" || code == "302"))
                  expected_method = "GET";
                if (code == "303" && method != "GET" && method != "HEAD")
                  expected_method = "GET";
                assert_equals(client.getResponseHeader("x-request-method"), expected_method)
                let expected_content_type = "application/x-pony";
                if (method == "POST" && (code == "301" || code == "302" || code == "303"))
                  expected_content_type = "NO";
                assert_equals(client.getResponseHeader("x-request-content-type"), expected_content_type);
                test.done()
              }
            })
          }
          client.open(method, `resources/redirect.py?location=${redirectLocation}&code=${code}`)
          client.setRequestHeader("Content-Type", "application/x-pony")
          client.send(null)
        })
      }
      redirect("301")
      redirect("302")
      redirect("303")
      redirect("307")
      redirect("301", "POST")
      redirect("302", "POST")
      redirect("303", "POST")
      redirect("307", "POST")
      redirect("301", "HEAD")
      redirect("302", "HEAD")
      redirect("303", "HEAD")
      redirect("307", "HEAD")

      let redirectedLocation = encodeURIComponent("redirect.py?location=content.py");
      redirect("301", "GET", redirectedLocation)
      redirect("302", "GET", redirectedLocation)
      redirect("303", "GET", redirectedLocation)
      redirect("307", "GET", redirectedLocation)
      redirect("301", "POST", redirectedLocation)
      redirect("302", "POST", redirectedLocation)
      redirect("303", "POST", redirectedLocation)
      redirect("307", "POST", redirectedLocation)
      redirect("301", "HEAD", redirectedLocation)
      redirect("302", "HEAD", redirectedLocation)
      redirect("303", "HEAD", redirectedLocation)
      redirect("307", "HEAD", redirectedLocation)
    </script>
  </body>
</html>
